comprobar si existe nombre de usuario o email en la base de datos al registrarse

+++ registro.php
@@ -5,6 +5,7 @@
     <title>Registrarse</title>
     <link href="css/style.css" rel="stylesheet" type="text/css">
     <link href="css/style-login.css" rel="stylesheet" type="text/css">
+    <script src="js/javascript.js"></script>
 </head>
 <body>
     <?php
@@ -19,25 +20,33 @@
         }
         else{
             if (filter_var($_POST['email'], FILTER_VALIDATE_EMAIL)) {
-                // Define $username and $password
-                $usuario = $_POST['usuario'];
-                $password = $_POST['password'];
-                $email = $_POST['email'];
-                $passw = SHA1($password);
-                $query = $pdo->prepare("INSERT INTO usuaris (id_usuari, usuari, password, email)VALUES(NULL, '$usuario', '$passw', '$email')");
-                $row = $query->execute();
-                if ($row == 1) {
-                    $headers = "MIME-Version: 1.0" . "\r\n";
-                    $headers .= "Content-type:text/html;charset=UTF-8" . "\r\n";
-                    $headers .= "From: Vota@proyectevota.org" . "\r\n";
-                    $mensaje = "Bienvenido " . $_POST['usuario'] . ".\n\n
-                                Tu registro a sido realizado con exito. \n 
-                                Tus datos de inicio de sesion son los siguientes:\n Usuario: " . $_POST['usuario'] . "\nContraseña: " . $_POST['password'] . "\n\n
-                                Gracias por registrarte.\nSaludos.";
-                    mail($_POST['email'], 'Registro con exito', $mensaje, $headers);
-                    header("location: index.php");
-                } else {
-                    echo "<script>error('No se ha podido insertar en la baase de datos!')</script>";
+                $comprobarDatos = $pdo->prepare("select count(email) existe from usuaris where email='" . $_POST['email'] . "' || usuari='" . $_POST['usuario'] . "'");
+                $comprobarDatos->execute();
+                if($comprobarDatos['existe']==0) {
+
+                    // Define $username and $password
+                    $usuario = $_POST['usuario'];
+                    $password = $_POST['password'];
+                    $email = $_POST['email'];
+                    $passw = SHA1($password);
+                    $query = $pdo->prepare("INSERT INTO usuaris (id_usuari, usuari, password, email)VALUES(NULL, '$usuario', '$passw', '$email')");
+                    $row = $query->execute();
+                    if ($row == 1) {
+                        $headers = "MIME-Version: 1.0" . "\r\n";
+                        $headers .= "Content-type:text/html;charset=UTF-8" . "\r\n";
+                        $headers .= "From: Vota@proyectevota.org" . "\r\n";
+                        $mensaje = "Bienvenido " . $_POST['usuario'] . ".\n\n
+                                    Tu registro a sido realizado con exito. \n 
+                                    Tus datos de inicio de sesion son los siguientes:\n Usuario: " . $_POST['usuario'] . "\nContraseña: " . $_POST['password'] . "\n\n
+                                    Gracias por registrarte.\nSaludos.";
+                        mail($_POST['email'], 'Registro con exito', $mensaje, $headers);
+                        header("location: index.php");
+                    } else {
+                        echo "<script>error('No se ha podido insertar en la baase de datos!')</script>";
+                    }
+                }
+                else{
+                    echo "<script>error('Este nombre de usuario o email ya estan registrados!')</script>";
                 }
             }
             else{
